+++ compiler/prog_data.m
@@ -2,7 +2,7 @@
 % vim: ft=mercury ts=4 sw=4 et
 %---------------------------------------------------------------------------%
 % Copyright (C) 1996-2012 The University of Melbourne.
-% Copyright (C) 2014-2022 The Mercury team.
+% Copyright (C) 2014-2024 The Mercury team.
 % This file may only be copied under the terms of the GNU General
 % Public License - see the file COPYING in the Mercury distribution.
 %---------------------------------------------------------------------------%
@@ -514,7 +514,7 @@ cons_id_is_const_struct(ConsId, ConstNum) :-
 
 :- type cons_exist_constraints
     --->    cons_exist_constraints(
-                % Neither list may be empty.
+                % cons_existq_tvars cannot be empty.
                 cons_existq_tvars   :: existq_tvars,
                 cons_constraints    :: list(prog_constraint),